requirements-dev.txt
numpy>=1.0
pytests